Kernel: Updated types to match those from stdint.h

+++ kernel/types.hpp
@@ -1,9 +1,11 @@
 #ifndef TYPES_HPP
 #define TYPES_HPP
 
+#include <stdint.h>
+
 using byte = unsigned char;
 using u32  = unsigned int;
-using u64  = unsigned long;
+using u64  = unsigned long long;
 using i32  = int;
 using i64  = long long;
 using f32  = float;
